requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
base58==2.1.1
bitarray==1.2.2
Bootstrap-Flask==2.0.2
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.3
cytoolz==0.11.2
eth-typing==2.3.0
eth-utils==1.10.0
Flask==2.1.2
Flask-SQLAlchemy==2.5.1
frozenlist==1.3.0
gunicorn==20.1.0
hexbytes==0.2.2
idna==3.3
importlib-metadata==4.11.4
importlib-resources==5.7.1
ipfshttpclient==0.8.0a2
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.5.1
lru-dict==1.1.7
MarkupSafe==2.1.1
multiaddr==0.0.9
multidict==6.0.2
mysql-connector-python==8.0.29
mysqlclient==2.1.0
netaddr==0.8.0
parsimonious==0.8.1
protobuf==3.20.1
py-solc-x==1.1.1
pycryptodome==3.14.1
pyrsistent==0.18.1
python-dotenv==0.20.0
requests==2.27.1
rlp==2.0.1
semantic-version==2.10.0
six==1.16.0
SQLAlchemy==1.4.36
toolz==0.11.2
urllib3==1.26.9
varint==1.0.2
web3==5.29.1
websockets==9.1
Werkzeug==2.1.2
WTForms==3.0.1
yarl==1.7.2
zipp==3.8.0
